Trait utils_db::traits::AsyncKVStorage

pub trait AsyncKVStorage {
    type Key: Serialize;
    type Value: Serialize;

    // Required methods
    fn get<'life0, 'async_trait>(
        &'life0 self,
        key: Self::Key
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Value>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set<'life0, 'async_trait>(
        &'life0 mut self,
        key: Self::Key,
        value: Self::Value
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Value>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn contains<'life0, 'async_trait>(
        &'life0 self,
        key: Self::Key
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove<'life0, 'async_trait>(
        &'life0 mut self,
        key: Self::Key
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Value>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn dump<'life0, 'async_trait>(
        &'life0 self,
        destination: String
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn iterate<'life0, 'async_trait>(
        &'life0 self,
        prefix: Self::Key,
        suffix_size: u32
    ) -> Pin<Box<dyn Future<Output = Result<StorageValueIterator<Self::Value>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn iterate_range<'life0, 'async_trait>(
        &'life0 self,
        start: Self::Key,
        end: Self::Key
    ) -> Pin<Box<dyn Future<Output = Result<StorageValueIterator<Self::Value>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch<'life0, 'async_trait>(
        &'life0 mut self,
        operations: Vec<BatchOperation<Self::Key, Self::Value>>,
        wait_for_write: bool
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn flush<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Associated Types

type Key: Serialize

type Value: Serialize

Required Methods

fn get<'life0, 'async_trait>( &'life0 self, key: Self::Key ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Value>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets a single database entry

fn set<'life0, 'async_trait>( &'life0 mut self, key: Self::Key, value: Self::Value ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Value>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sets a single database entry

fn contains<'life0, 'async_trait>( &'life0 self, key: Self::Key ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns true if database contains an entry that matches the given key

fn remove<'life0, 'async_trait>( &'life0 mut self, key: Self::Key ) -> Pin<Box<dyn Future<Output = Result<Option<Self::Value>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Removes the database entry that matches the given key

fn dump<'life0, 'async_trait>( &'life0 self, destination: String ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Dumps the contents of the database into a file

fn iterate<'life0, 'async_trait>( &'life0 self, prefix: Self::Key, suffix_size: u32 ) -> Pin<Box<dyn Future<Output = Result<StorageValueIterator<Self::Value>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns an iterator that yields all database entries whose key matches the given prefix and the length of the suffix. Does not match shorter or longer suffixes, even though they have the right suffix.

fn iterate_range<'life0, 'async_trait>( &'life0 self, start: Self::Key, end: Self::Key ) -> Pin<Box<dyn Future<Output = Result<StorageValueIterator<Self::Value>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns an iterator that yields all database entries whose is in the interval from start (inclusive) and end (inclusive).

fn batch<'life0, 'async_trait>( &'life0 mut self, operations: Vec<BatchOperation<Self::Key, Self::Value>>, wait_for_write: bool ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Constructs batch query

fn flush<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Flushes data to disk